Resize input instead of actipy resample
