+++ src/actinet/prepare.py
@@ -1,13 +1,12 @@
 import numpy as np
 import pandas as pd
-from scipy.interpolate import interp1d
 import os
 from tqdm.auto import tqdm
 from joblib import Parallel, delayed
 import warnings
 import actipy
 
-from actinet.utils.utils import is_good_window
+from actinet.utils.utils import is_good_window, resize
 
 
 def load_data(datafile, sample_rate=100, annot_type=str):
@@ -37,26 +36,6 @@ def load_data(datafile, sample_rate=100, annot_type=str):
 
     data, _ = actipy.process(data, sample_rate, verbose=False)
     return data
-
-
-def resize(x, length, axis=1):
-    """
-    Resize the temporal length of the data using linear interpolation.
-
-    Args:
-        x (ndarray): Data to be resized.
-        length (int): New length of the data.
-        axis (int): Axis along which to perform the interpolation. Defaults to 1.
-
-    Returns:
-        ndarray: Resized data.
-
-    """
-    length_orig = x.shape[axis]
-    t_orig = np.linspace(0, 1, length_orig, endpoint=True)
-    t_new = np.linspace(0, 1, length, endpoint=True)
-    x = interp1d(t_orig, x, kind="linear", axis=axis, assume_sorted=True)(t_new)
-    return x
 
 
 def make_windows(
